Added int64 support to the script engine
Added support for signed 64 bit integer value computation into the script engine.
This is required because newer official features require support for bigger numbers inside the scripts.

This also cleans up a lot of messy code and fixes a few issues like script stack corruptions.

+++ src/map/tax.cpp
@@ -32,7 +32,7 @@ const std::string TaxDatabase::getDefaultLocation() {
 * @return count of successfully parsed rows
 */
 uint64 TaxDatabase::parseBodyNode(const YAML::Node &node) {
-	int type = 0;
+	int64 type = 0;
 	std::string type_str;
 
 	if (!this->nodeExists(node, "Type")) {
